Move mocks to subdirectory
