+++ tests/Mock/MockConsoleOutput.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types = 1);
 
-namespace MountHolyoke\JorgeTests;
+namespace MountHolyoke\JorgeTests\Mock;
 
 use Symfony\Component\Console\Output\ConsoleOutput;
